--- backend/ErrorCentral.WebAPI/ErrorCentral.AppDomain/Interfaces/IAuthorizationService.cs
@@ -1,9 +1,0 @@
-using ErrorCentral.AppDomain.Models;
-
-namespace ErrorCentral.AppDomain.Interfaces
-{
-    public interface IAuthorizationService
-    {
-        BaseResult<IUser> Authorize(LoginUser loginUser);
-    }
-}
