+++ backend/ErrorCentral.WebAPI/ErrorCentral.AppDomain/Interfaces/IUserService.cs
@@ -7,6 +7,7 @@ namespace ErrorCentral.AppDomain.Interfaces
 {
     public interface IUserService
     {
+        BaseResult<IUser> Authorize(LoginUser loginUser);
         List<User> Get();
         User GetByEmail(string Email);
         User Save(User user);
